Socket Backend Client Project Update
This project could become the main project for our CSCI 355 FTP client
project, but it's separate for now. This is just a simple backend
program that creates sockets and connects to a specified server. This is
different than the previous program under this repo because I redesigned
the code & fixed some bugs.